Update README.md
# Camagru

Camagru is an instagram-like web app that has webcam integration and basic image editing capabilities. You can capture an image with webcam or upload an image and add stickers on top of the image. Stickers can be positioned freely with mouse, and the sticker's size and opacity can be modified. Images can be saved as drafts or posted to be displayed in image feed to other users. The feed features infinite pagination and the posts have like and comment capabilities. User settings can be modified in control panel.

The app has been containerized with Docker and deployed on basic Linux/NGINX server with traffic routed through Cloudflare. The app is available for testing at (most functionality requires registration with email confirmation):

https://jukkay.dev/


## Technologies
- PHP
- Vanilla JavaScript
- HTML
- CSS
- Bulma CSS
- MariaDB(MySQL)
- Fetch API
- Docker
- NGINX
- Linux


## Disclaimer

Camagru is one of the projects at 42 school's web development branch, so the idea isn't mine. But it was a solo project, so 100% of the work was done by me. If you're interested you can find the project assignment [here](./camagru.en.pdf)


## Usage

You have to options:

1. Just check out the deployed version at https://jukkay.dev/
2. Or clone the repository and use the docker-compose file to whip up the containers.